refactor: enhance Firebase service and HTTP function structure
- Updated Firebase class to use private properties for Firestore, Auth, Storage, and App instances.
- Added optional parameters for Firestore and Auth methods to support multi-tenancy.
- Refactored createFunction to include parameters for isolating controllers and removing controller prefixes.
- Enhanced createFirebaseHttpsV1 and createFirebaseHttpsV2 to accept runtime options and controller management parameters.
- Removed deprecated FirestoreTriggerV1Options interface and replaced it with RuntimeOptions.
- Improved error handling in getModule function.
- Updated tsconfig.json to enable experimental decorators and emit design-type metadata.

+++ src/httpFunction/create-function.ts
@@ -7,15 +7,26 @@ import { deleteControllerPrefix } from './delete-controller-prefix';
 
 const loadedFunctions: { [key: string]: any } = {};
 
-export async function createFunction(module: any, expressServer: Express): Promise<INestApplication> {
+export async function createFunction(
+  module: any,
+  expressServer: Express,
+  isolateControllers: boolean = true,
+  removeControllerPrefix: boolean = true
+): Promise<INestApplication> {
   const moduleName = module.constructor.name;
   if (loadedFunctions[moduleName]) {
     return loadedFunctions[moduleName];
   }
-  deleteImportedControllers(module);
-  deleteControllerPrefix(module);
-  const app: INestApplication = await NestFactory.create(module, new ExpressAdapter(expressServer));
 
+  if (isolateControllers) {
+    deleteImportedControllers(module);
+  }
+
+  if (removeControllerPrefix) {
+    deleteControllerPrefix(module);
+  }
+
+  const app: INestApplication = await NestFactory.create(module, new ExpressAdapter(expressServer));
   app.useGlobalPipes(new ValidationPipe());
   app.enableCors({
     methods: 'GET,HEAD,PUT,PATCH,POST,DELETE',
@@ -23,5 +34,6 @@ export async function createFunction(module: any, expressServer: Express): Promi
   });
   const appInit = await app.init();
   loadedFunctions[moduleName] = appInit;
+
   return appInit;
 }
